feat: Split CSS folder into separate directories
Css splitted for components.Components subfolders created
