+++ material/src/components/emailField/emailField.css
@@ -1,0 +1,37 @@
+#email {
+    border: grey solid 1px;
+    border-radius: 4px;
+   }
+   
+   #email:hover {
+     border: white solid 3px;
+     transition: 0ms;
+   }
+  
+   #elabel {
+    position: absolute;
+    left: 10px;
+    top: 25px;
+    transition: top 0.3s ease, left 0.3s ease, font-size 0.3s ease;
+  }
+  
+  #email:focus + #elabel
+  {
+    top: -5px;
+    left: 10px;
+    font-size: 14px;
+    color: #5fa8d3;
+    background-color: #34495e;
+    padding-left: 5px;
+    padding-right: 5px;
+  }
+
+  #elabel.filled {
+    top: -5px; 
+    left: 13; 
+    font-size: 14px; 
+    background-color: #34495e;
+    padding-left: 5px;
+    padding-right: 5px;
+  }
+
