+++ material/src/components/passwordField/passwordField.css
@@ -1,0 +1,28 @@
+#plabel {
+    position: absolute;
+    left: 10px;
+    top: 22px;
+    transition: top 0.3s ease, left 0.3s ease, font-size 0.3s ease;
+  }
+
+  #plabel.filled {
+    top: 0;
+    left: 0;
+    font-size: 12px; 
+  }
+  
+  #password:focus + #plabel {
+    top: 0;
+    left: 0;
+    font-size: 12px;
+  }
+  
+  #password {
+    background-color: #3f5770;
+    border-radius: 4px;
+  }
+
+  #password:hover {
+    background-color: #557496;
+    transition: background-color 0.3s ease;
+  }
