+++ material/src/components/passwordField/passwordField.jsx
@@ -1,6 +1,6 @@
 import React, { useState } from 'react';
-import { InputField } from '../scripts/inputFocus';
-import { PasswordToggle } from '../scripts/passwordToggle';
+import { InputField } from '../../scripts/inputFocus';
+import { PasswordToggle } from '../../scripts/passwordToggle';
 
 export const PasswordField = () => {
   const [passwordVisible, setPasswordVisible] = useState(false);
