+++ material/src/components/submitButton/submitButton.css
@@ -1,0 +1,39 @@
+#submit-button {
+    position: relative;
+    overflow: hidden;
+    background-color: #90caf9;
+    transition: background-color 0.3s;
+    border: none;
+    color: black;
+    padding: 15px 32px;
+    text-align: center;
+    display: inline-block;
+    font-size: 16px;
+    margin: 4px 2px;
+    cursor: pointer;
+    border-radius: 8px;
+  }
+  
+  #submit-button:after {
+    content: "";
+    background: rgba(255, 255, 255, 0.3);
+    display: block;
+    position: absolute;
+    padding-top: 300%;
+    padding-left: 350%;
+    margin-left: -40px !important;
+    margin-top: -120%;
+    opacity: 0;
+    transition: all 0.8s
+  }
+  
+  #submit-button:active:after {
+    padding: 0;
+    margin: 0;
+    opacity: 1;
+    transition: 0s
+  }
+  
+  #submit-button:hover {
+    background-color: #3d8ccc;
+  }
