+++ material/src/main.css
@@ -1,0 +1,45 @@
+body {
+    display: flex;
+    background-color: #34495e;
+    justify-content: center;
+    align-items: center;
+    height: 100vh;
+    font-family: Arial, sans-serif;
+  }
+  
+  .card {
+    padding: 40px;
+    border-radius: 10px;
+    box-shadow: 0px 0px 10px 0px rgba(0, 0, 0, 0.1);
+  }
+  
+  .input-field {
+    position: relative;
+    width: 200px;
+    margin-top : 20px;
+    margin-bottom: 20px;
+  }
+  
+  .input-field label {
+    position: absolute;
+    top: 40%;
+    transition: 0.3s;
+    pointer-events: none;
+    color: white;
+    margin-bottom: 20px;
+  }
+  
+  .input-field input {
+    width: 100%;
+    height: 50px;
+    color: white;
+    border: none;
+    outline: none;
+    background-color: #34495e;
+    border-bottom:2px solid white;
+  }
+  
+  .input-field input:hover {
+    border-bottom: 4px solid white;
+    transition: 0s;
+  }
