--- material/src/styles.css
@@ -1,170 +1,0 @@
-body {
-    display: flex;
-    background-color: #34495e;
-    justify-content: center;
-    align-items: center;
-    height: 100vh;
-    font-family: Arial, sans-serif;
-  }
-  
-  .card {
-    padding: 40px;
-    border-radius: 10px;
-    box-shadow: 0px 0px 10px 0px rgba(0, 0, 0, 0.1);
-  }
-  
-  .input-field {
-    position: relative;
-    width: 200px;
-    margin-top : 20px;
-    margin-bottom: 20px;
-  }
-  
-  .input-field label {
-    position: absolute;
-    top: 40%;
-    transition: 0.3s;
-    pointer-events: none;
-    color: white;
-    margin-bottom: 20px;
-  }
-  
-  .input-field input {
-    width: 100%;
-    height: 50px;
-    color: white;
-    border: none;
-    outline: none;
-    background-color: #34495e;
-    border-bottom:2px solid white;
-  }
-  
-  .input-field input:hover {
-    border-bottom: 4px solid white;
-    transition: 0s;
-  }
-  
-  #submit-button {
-    position: relative;
-    overflow: hidden;
-    background-color: #90caf9;
-    transition: background-color 0.3s;
-    border: none;
-    color: black;
-    padding: 15px 32px;
-    text-align: center;
-    display: inline-block;
-    font-size: 16px;
-    margin: 4px 2px;
-    cursor: pointer;
-    border-radius: 8px;
-  }
-  
-  #submit-button:after {
-    content: "";
-    background: rgba(255, 255, 255, 0.3);
-    display: block;
-    position: absolute;
-    padding-top: 300%;
-    padding-left: 350%;
-    margin-left: -40px !important;
-    margin-top: -120%;
-    opacity: 0;
-    transition: all 0.8s
-  }
-  
-  #submit-button:active:after {
-    padding: 0;
-    margin: 0;
-    opacity: 1;
-    transition: 0s
-  }
-  
-  #submit-button:hover {
-    background-color: #3d8ccc;
-  }
-  
-  #email {
-    border: grey solid 1px;
-    border-radius: 4px;
-   }
-   
-   #email:hover {
-     border: white solid 3px;
-     transition: 0ms;
-   }
-  
-   #elabel {
-    position: absolute;
-    left: 10px;
-    top: 25px;
-    transition: top 0.3s ease, left 0.3s ease, font-size 0.3s ease;
-  }
-  
-  #email:focus + #elabel
-  {
-    top: -5px;
-    left: 10px;
-    font-size: 14px;
-    color: #5fa8d3;
-    background-color: #34495e;
-    padding-left: 5px;
-    padding-right: 5px;
-  }
-
-  #elabel.filled {
-    top: -5px; 
-    left: 13; 
-    font-size: 14px; 
-    background-color: #34495e;
-    padding-left: 5px;
-    padding-right: 5px;
-  }
-  
-   #plabel {
-    position: absolute;
-    left: 10px;
-    top: 22px;
-    transition: top 0.3s ease, left 0.3s ease, font-size 0.3s ease;
-  }
-
-  #plabel.filled {
-    top: 0;
-    left: 0;
-    font-size: 12px; 
-  }
-  
-  #password:focus + #plabel {
-    top: 0;
-    left: 0;
-    font-size: 12px;
-  }
-  
-  #password {
-    background-color: #3f5770;
-    border-radius: 4px;
-  }
-
-  #password:hover {
-    background-color: #557496;
-    transition: background-color 0.3s ease;
-  }
-  
-  #ulabel {
-    position: absolute;
-    left: 10px;
-    top: 25px;
-    transition: top 0.3s ease, left 0.3s ease, font-size 0.3s ease;
-  }
-  
-  #ulabel.filled {
-    top: 0; 
-    left: 0; 
-    font-size: 12px; 
-  }
-
-  #username:focus + #ulabel {
-    top: 0;
-    left: 0;
-    font-size: 12px;
-  }
